@@ -1,13 +1,12 @@
-﻿namespace Tellurian.Trains.Modules
+﻿namespace Tellurian.Trains.Modules;
+
+public interface ILocoConfigurationStore
 {
-    public interface ILocoConfigurationStore
-    {
-        LocoConfiguration? TryLoad(int address);
-        void Save(LocoConfiguration configuration);
-    }
-    public interface IDecoderService
-    {
-        LocoConfiguration? Read();
-        void Write(LocoConfiguration configuration);
-    }
+    LocoConfiguration? TryLoad(int address);
+    void Save(LocoConfiguration configuration);
+}
+public interface IDecoderService
+{
+    LocoConfiguration? Read();
+    void Write(LocoConfiguration configuration);
 }
